Examples and user models reorganized
