+++ dq0sdk/data/__init__.py
@@ -5,7 +5,7 @@ This package contains the data connector abstract classes and
 implementing subclasses.
 """
 
-from . import census_data, cifar10, csv, user, utils
+from . import census, cifar, csv, user, utils
 from .connector import Connector
 from .source import Source
 
@@ -15,6 +15,6 @@ __all__ = [
     'census',
     'csv',
     'user',
-    'cifar10',
+    'cifar',
     'utils'
 ]
