+++ dq0sdk/models/tf/__init__.py
@@ -5,7 +5,6 @@ This package contains the tensorflow models subclassing the abstract
 model base class.
 """
 
-from .image_cnn import ImageCNN
 from .neural_network import NeuralNetwork
 from .neural_network_yaml import NeuralNetworkYaml
 from .tf_hub import TFHub
@@ -13,6 +12,5 @@ from .tf_hub import TFHub
 __all__ = [
     'NeuralNetwork',
     'NeuralNetworkYaml',
-    'TFHub',
-    'ImageCNN'
+    'TFHub'
 ]
